Fix attribute merging for python builder
Previously, `extraDrvAttrs` would not be applied to the python builder,
because it's merged with the derivation attributes **after** the
`pkgs.python3Packages.buildPythonApplication` function has been applied
to it, resulting in an undesirable behaviour.

This can be tested by adding:

```nix
extraDrvAttrs = {
  preBuild = ''
    exit 1
  '';
};
```

And examining whether the build exits or not.